reorganizing NavbarDrawer components and styling
